zap/zap-scans/src/requirements.txt
aiohttp==3.7.3
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==20.3.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
codedx-api @ git+https://github.com/broadinstitute/dsp-appsec-codedx-api-client-python.git@b0bea14ef9f39df4e3abfb187ac7d2478ef8c865
cryptography==2.1.4
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-pubsub==2.2.0
googleapis-common-protos==1.52.0
grpc-google-iam-v1==0.12.3
grpcio==1.35.0
httplib2==0.18.1
idna==2.10
idna-ssl==1.1.0
keyring==10.6.0
keyrings.alt==3.0
libcst==0.3.16
multidict==5.1.0
mypy-extensions==0.4.3
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-owasp-zap-v2.4==0.0.18
pytz==2020.5
pyxdg==0.25
PyYAML==5.4.1
requests==2.25.1
rsa==4.7
SecretStorage==2.3.1
six==1.15.0
slackclient==2.9.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
yarl==1.6.3
